migration of motor interface def
